# perf: 提升 to_string 转换性能