标题：perf: 提升 to_string 转换性能

动机
- 目标场景：`to_string` 在 256 位整数上仍落后于 Boost（600ns vs 418ns）。
- 基线：旧实现使用 1e9 基数，多次除法，成为主要开销。

优化思路
- 算法/路径：改用 1e19 基数减少迭代次数，首块与后续块均手写十进制转换并预留缓冲。
- 位宽专用化：无。
- 可维护性：逻辑集中在通用路径，可读性可接受。

正确性
- 语义不变：十进制字符串与符号处理未变。
- 覆盖测试：`make test` 通过。

基准对比（必填）
- 环境：x86_64，GCC，`-O3 -DNDEBUG`。
- 方法：`BENCH_ARGS=--benchmark_filter=ToString make bench-compare`
- 数据：变更前 600ns（gint）/418ns（Boost），变更后 263ns（gint）/302ns（Boost）。
- 总结：新实现较旧版本约快 56%，并领先 Boost 约 13%。

回归与风险
- 平台差异：依赖 128 位内建除法，Clang/GCC 均支持。
- 代码尺寸与可读性：增加少量循环和缓冲处理，可维护性可接受。

清单（提交前请逐项勾选）
- [x] 分支名为英文（PR 指南要求）
- [x] 已运行 clang-format（使用仓库根目录 `.clang-format`）
- [x] 新增/更新了 `tests/` 覆盖关键路径，`make test` 通过
- [x] 已提供“变更前/后”基准数据（必填）
- [x] 若更改基准用例，已更新 `bench/` 与 `docs/BENCHMARKS.md`
- [x] 保持 C++11 兼容与 Header-only 特性
- [x] 保持严格位宽与原生类型互操作语义不变

------
https://chatgpt.com/codex/tasks/task_e_68b6fc5016f083299af14a4a8307b79d